Module 5 Final Project
Module 5 Final Project: Update 3 - Classes/Test Classes
These classes are going to be the skeletal structure of my Java app that I am going to continue focusing on. The app has been tested. It reminds the user when their birthday is in order to prepare for it. If the date is the exact date of their birthday, the program says it is their birthday, go celebrate!