JavaFX Assignment 3
This is a program that displays a clock and sets the time with the input from three text fields I tested the hours, minutes and seconds, and they accurately depict the picture shown in Figure 16.38b. The ClockPane from Listing 14.21 is used for this exercise. The clock is resized to the center of the pane.